Add files via upload
Enter marks from keyboard for 5 different subjects and calucate the percentage. The student gets a division as per the following rules. :
The student gets a division as per the following rules: Percentage above or equal to 60-- First Division, 
Percentage above or equal to 50 and 59 -- Second Division,
Percentage above or equal to 40 and 49 -- Third Division,
Percentage less than 40 -- Fail .